fix ephemeris error for multisector PCs
Previously, I took an average of all the fitted parameters (t0, period,
duration, a/Rstar). This is wrong for t0.

This commit instead fits a line to the multiple ephemeris points, and
uses the errors from the least-square covariance matrix if they are
finite. If they are infinite (two points / sectors available), it takes
the most recent ephemeris.