allvariable reports: patched a memory leak
Somewhere in this code, there's a memory leak. It's probably linked
somehow to the implicit requests module calls in the DSS or GAIA
querying. It means that running drivers.do_allvariable_report_making.py
for >~1 hour continuously (i.e., about 50 reports) leads to a memory
overflow.

The easiest way to fix it? Literally just kill the driver, and restart
it. This is the approach taken in drivers/run_allvar_driver.sh