PCA / reformat_header / get_dtrvecs: speed-up x4
Previous runs showed that for a single light curve, a call to
reformat the HLSP headers (given the {source_id}_llc.fits files)
took ~6 seconds on average for a single light curve.

This is a throttle for large-scale light curve production.

The reason came from `get_dtrvecs`, in which the eigenvector smoothing
was being done as a post-processing step to the PCA eigenvectors  (by
`eigvec_smooth_fn`).  This smooth was being run on all 200 components of
the eigenvectors for each aperture.  This process took ~1.5 seconds for
each aperture, so 4.5 of the 6 seconds.

The above process really does not need to be repeated for every light
curve in a sector -- it can be done at the sector/camera/ccd tuple
level, as is already the case for the PCA eigenvector construction.
This yields the speed-up mentioned in the tagline.

As an example use case, the <200pc, G<13 sample has ~3e5 stars and light
curves required.   The reformatting time for such a sample drops from
~16 days to ~4 days this way.

Note that this didn't crop up in earlier Cycle1/2 reductions, because
the eigenvector smoothing was not being done in those.