prepare_pca: address nan-filled template edge case
For the cycle 3 and 4 data, we adopted a less aggressive approach toward
"pre"-nan-masking garbage times with heavy scattered light.

This led to more nan-filled light curves, where NaNs came from fiphot
rather than from our initial masking.

This matters for the template creation process behind PCA, because PCA
by default does not play well with nans.

In fact, sometimes, only a small fraction of the "template" light curves
were nan free!

This change is designed to address such cases.

The approach that is now taken (in such NaN-filled cases) is to
interpolate (linearly in _cadence_) over the NaN gaps in the template
light curve vectors being fed into the PCA.

This yields a set of template vectors that is large enough, and that has
size matching the number of *times* in the sector, for PCA to correctly
run.

The previous behavior for this case was for the PCA to crash.  So, while
the interpolation will slightly affect the PCA weights, there was no
alternative previously implemented!