[WIP] nuance: switch kernels to "rotation"
...meaning the kernel of 2 SHOs, plus two damped exponentials.

This thing is still having what I view as overfitting problems.

I need to mull it over more, but it seems to me that if your
max-likelihood hyperparameters lead to overfitting, then you're always
going to have reduced SNR in your fitted light curves.

Also, the billing of "simultaneous GP + BLS period search" suggests
to the naive practioner that one is somehow able to marginalize over the
GP's hyperparameters, so that when computing the periodogram they do not
matter.  That is not the case.  Some hyperparameters will give a higher
SNR detection of a given planet signal than others.  So: how can one be
sure that one is using the GP hyperparameters that in fact maximize the
SNR for planet detection?

(This is reminiscent of the problem in "detrending -> BLS" style
approaches of setting e.g. the window slider lengths for single-light
curve detrenders)