[WIP] nuance: add run_iterative_nuance
NB: still need to add an optional exit condition, rather than just
running max_iter iterations as is currently implemented.

This condition depends in an important way on what the meaning of "SNR"
is in the periodograms that nuance is returning.

As-implemented, seems to work fine on TOI-837; Kepler-1627 is running.